Rubocopped this repo. Went from 200+ errors to 1 "error".
There may be a few bugs for longer files like tic-tac-toe but i plan to do an
extensive refactor on it anyway. The goal is to get my code to follow best
practices.

FYI - The only 'error' remaining is the 'method too' long for `bubble_sort_2`
since it's 11 lines long. I think this is acceptable since the code is lean.